test: Revenue 테스트 분리
